backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
geoalchemy2==0.15.2
alembic==1.14.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
anthropic>=0.84.0
httpx==0.28.1
sse-starlette==2.2.1
tzdata>=2024.1
jieba>=0.42.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.9
twilio>=9.0
line-bot-sdk>=3.0
